usability test issues fixed +general layout updates

+++ lib/screens/edit_provider_account_screen.dart
@@ -1,5 +1,6 @@
 import 'dart:io' show Platform;
 
+import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:genchi_app/components/drop_down_services.dart';
@@ -38,6 +39,7 @@ class EditProviderAccountScreen extends StatefulWidget {
 
 class _EditProviderAccountScreenState extends State<EditProviderAccountScreen> {
   FirestoreAPIService firestoreAPI = FirestoreAPIService();
+  FirebaseAnalytics analytics = FirebaseAnalytics();
 
   TextEditingController nameTextController = TextEditingController();
   TextEditingController bioTextController = TextEditingController();
@@ -52,7 +54,6 @@ class _EditProviderAccountScreenState extends State<EditProviderAccountScreen> {
   bool showSpinner = false;
   String service;
   String pricing;
-
   bool changesMade = false;
 
 
@@ -128,6 +129,7 @@ class _EditProviderAccountScreenState extends State<EditProviderAccountScreen> {
                   color: Colors.black,
                 ),
                 onPressed: () async {
+                  analytics.logEvent(name: 'provider_top_save_changes_button_pressed');
                   setState(() {
                     showSpinner = true;
                   });
@@ -313,6 +315,7 @@ class _EditProviderAccountScreenState extends State<EditProviderAccountScreen> {
                   buttonTitle: 'Save changes',
                   buttonColor: Color(kGenchiGreen),
                   onPressed: ()async{
+                    analytics.logEvent(name: 'provider_bottom_save_changes_button_pressed');
                     setState(() {
                       showSpinner = true;
                     });
@@ -342,6 +345,7 @@ class _EditProviderAccountScreenState extends State<EditProviderAccountScreen> {
                   buttonColor: Color(kGenchiBlue),
                   elevation: false,
                   onPressed: () async {
+
                     ///Update the provider before deleting
 
                     await providerService
@@ -352,45 +356,30 @@ class _EditProviderAccountScreenState extends State<EditProviderAccountScreen> {
                         Provider.of<ProviderService>(context, listen: false)
                             .currentProvider;
 
-                    Platform.isIOS
-                        ? showAlertIOS(
-                            context: context,
-                            actionFunction: () async {
-                              setState(() => showSpinner = true);
+                    bool delete = await showYesNoAlert(context: context, title: "Delete Provider Account?");
 
-                              await firestoreAPI.deleteProvider(
-                                  provider: providerUser);
-                              await authProvider.updateCurrentUserData();
-                              changesMade = false;
-                              setState(() => showSpinner = false);
+                    if(delete) {
+                      setState(() {
+                        showSpinner = true;
+                      });
 
-                              Navigator.pushNamedAndRemoveUntil(
-                                  context,
-                                  HomeScreen.id,
-                                  (Route<dynamic> route) => false,
-                                  arguments:
-                                      HomeScreenArguments(startingIndex: 3));
-                            },
-                            alertMessage: 'Delete Account')
-                        : showAlertAndroid(
-                            context: context,
-                            actionFunction: () async {
-                              setState(() => showSpinner = true);
+                      ///Log event in firebase
+                      await analytics.logEvent(name: 'provider_account_deleted');
 
-                              await firestoreAPI.deleteProvider(
-                                  provider: providerUser);
-                              await authProvider.updateCurrentUserData();
-                              changesMade = false;
-                              setState(() => showSpinner = false);
+                      await firestoreAPI.deleteProvider(
+                          provider: providerUser);
+                      await authProvider.updateCurrentUserData();
+                      changesMade = false;
+                      setState(() => showSpinner = false);
 
-                              Navigator.pushNamedAndRemoveUntil(
-                                  context,
-                                  HomeScreen.id,
-                                  (Route<dynamic> route) => false,
-                                  arguments:
-                                      HomeScreenArguments(startingIndex: 3));
-                            },
-                            alertMessage: "Delete Account");
+                      Navigator.pushNamedAndRemoveUntil(
+                          context,
+                          HomeScreen.id,
+                              (Route<dynamic> route) => false,
+                          arguments:
+                          HomeScreenArguments(startingIndex: 3));
+
+                    }
                   },
                 ),
               ],
